backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
boto3==1.34.34
langchain
langchain-community
langchain-openai
langchain-text-splitters
openai
numpy<2.0.0
faiss-cpu==1.7.4
pypdf